Hotfix: Update css path and convert api file to ts
